rename pillar to module, add README documentation
- Rename AnalyticsPillar.js to AnalyticsModule.js
- Rename AnalysisPillar.js to AnalysisModule.js
- Rename PreviewPillar.js to PreviewModule.js
- Update all imports and references across codebase
- Rename .pillar-container CSS class to .module-container
- Add comprehensive README with project info, collaborators, and setup instructions
- Add Sharjeel Shaikh and University of Potsdam as collaborators
