+++ frontend/app/analytics/loading.tsx
@@ -4,7 +4,7 @@ import Loader from '@/components/Loader'
 
 const Loading = () => {
   return (
-    <Loader name={"Analytics Pillar"}/>
+    <Loader name={"Analytics Module"}/>
   )
 }
