+++ frontend/app/globals.css
@@ -88,7 +88,7 @@ body {
   transition: background-color 0.3s, color 0.3s;
 }
 
-.pillar-container {
+.module-container {
   overflow: hidden; 
   height: 90vh; 
 }
