+++ frontend/app/preview/[video_name]/loading.tsx
@@ -3,7 +3,7 @@ import Loader from '@/components/Loader'
 const Loading = () => {
   console.log("Loading render")
   return (
-    <Loader name={"Preview Pillar"}/>
+    <Loader name={"Preview Module"}/>
   )
 }
