+++ frontend/tsconfig.json
@@ -21,7 +21,7 @@
     "paths": {
       "@/*": ["./*"]
     },
-    "checkJs": false // temp_solution -- preview pillar missing types
+    "checkJs": false // temp_solution -- preview module missing types
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/preview/[video_name]/page.js"],
   "exclude": ["node_modules"],
